Change 1: I added multiple states to the game. As of now, there is a start state(not completely working, but accepts key input), and a gameplay state(which the game switches to when 'space is pressed in the start state). The gameplay works normally.
Change 2:
I changed the init.js file such that there will be no animation when the
window opens. This is to make it easier to set up the global variables
storing the width and height of the window. I did this to avoid
accessing the DOM instide the game_animate() function, for it made the
game stuttery and reduced its framerate.